Added --test-all flag
Some changes needed to be made: Benchmarks are now only done for all
the application processing, so no fine grained benchmarks. We should
add them in the future, but we need to modify the way we are doing
benchmarking. We should probably also use other libraries for
benchmarks.

Still some comments should be improved. Cyclomatic complexity of the
main() function has increased, though, and that should be fixed.