Upgraded dependencies.
Biggest upgrade was Regex 0.1 -> Regex 0.2. This makes SUPER ready
for Regex 1.0 that will came in the near future. Some changes in
the match structures had to be done.

Handlebars had to be updated to 0.24 but I'm having problems
compiling it. I guess I will update it when I fix those problems.