Automated Linux packaging and deployment (#164)
This will enable deploying Linux packages for Ubuntu, CentOS, Fedora and Debian in each tag. Not only that, packages will be built in each commit, to make sure nothing breaks.

It also adds a new place where the release version must be updated with a new release, maybe that could someday be automated. We now need to check the `Cargo.toml` file, the `travis.yml` file and the `rpmbuild/super.spec` file.

As an extra, given that I was working in this branch, I implemented a fix for #159, so all three, #40, #149 and #159 are fixed now.